test: add mock-accounts.ts
