+++ ui/selectors/accounts.test.ts
@@ -1,63 +1,10 @@
-import { KeyringTypes } from '@metamask/keyring-controller';
 import {
-  InternalAccount,
-  EthAccountType,
-  BtcMethod,
-  BtcAccountType,
-} from '@metamask/keyring-api';
-import {
-  ETH_EOA_METHODS,
-  ETH_4337_METHODS,
-} from '../../shared/constants/eth-methods';
+  MOCK_ACCOUNTS,
+  MOCK_ACCOUNT_EOA,
+  MOCK_ACCOUNT_ERC4337,
+  MOCK_ACCOUNT_BIP122_P2WPKH,
+} from '../../test/data/mock-accounts';
 import { AccountsState, isSelectedInternalAccountEth } from './accounts';
-
-export const MOCK_ACCOUNT_EOA: InternalAccount = {
-  id: '4974fc00-c0fb-4a18-8535-8407ec6d1952',
-  address: '0x123',
-  options: {},
-  methods: ETH_EOA_METHODS,
-  type: EthAccountType.Eoa,
-  metadata: {
-    name: 'Account 1',
-    keyring: { type: KeyringTypes.hd },
-    importTime: 1691565967600,
-    lastSelected: 1691565967656,
-  },
-};
-
-export const MOCK_ACCOUNT_ERC4337: InternalAccount = {
-  id: '4d5921f2-2022-44ce-a84f-9f6a0f142a5c',
-  address: '0x123',
-  options: {},
-  methods: ETH_EOA_METHODS.concat(ETH_4337_METHODS),
-  type: EthAccountType.Erc4337,
-  metadata: {
-    name: 'Account 2',
-    keyring: { type: KeyringTypes.snap },
-    importTime: 1691565967600,
-    lastSelected: 1691565967656,
-  },
-};
-
-export const MOCK_ACCOUNT_BIP122_P2WPKH: InternalAccount = {
-  id: 'ae247df6-3911-47f7-9e36-28e6a7d96078',
-  address: 'bc1qaabb',
-  options: {},
-  methods: [BtcMethod.SendMany],
-  type: BtcAccountType.P2wpkh,
-  metadata: {
-    name: 'Bitcoin Account',
-    keyring: { type: KeyringTypes.snap },
-    importTime: 1691565967600,
-    lastSelected: 1955565967656,
-  },
-};
-
-export const MOCK_ACCOUNTS = {
-  [MOCK_ACCOUNT_EOA.id]: MOCK_ACCOUNT_EOA,
-  [MOCK_ACCOUNT_ERC4337.id]: MOCK_ACCOUNT_ERC4337,
-  [MOCK_ACCOUNT_BIP122_P2WPKH.id]: MOCK_ACCOUNT_BIP122_P2WPKH,
-};
 
 const MOCK_STATE: AccountsState = {
   metamask: {
